backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
google-generativeai
python-multipart
httpx
pytest
groq
PyMuPDF
fpdf2
supabase==2.10.0
python-dotenv==1.0.1